Copy CNAME if present in Make target

+++ Makefile
@@ -4,7 +4,7 @@ html:
 	# Check for ipynb files in source (should all be .Rmd).
 	if compgen -G "*.ipynb" 2> /dev/null; then (echo "ipynb files" && exit 1); fi
 	jupyter-book build -W .
-	cp CNAME $(BUILD_DIR)
+	if [ -e "CNAME" ]; then cp CNAME $(BUILD_DIR); fi
 
 clean:
 	rm -rf _build
